requirements.txt.txt
streamlit 
selenium 
beautifulsoup4 
aiohttp 
pandas 
pillow 
requests